DAO complete

+++ src/main/java/by/epam/javawebtraining/bean/Entity.java
@@ -3,20 +3,20 @@ package by.epam.javawebtraining.bean;
 import by.epam.javawebtraining.dao.daointerface.IDdefinition;
 
 public class Entity {
-    protected int id;
+    protected long id;
 
     protected Entity() {
     }
 
-    protected Entity(int id) {
+    protected Entity(long id) {
         this.id = id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 }
